requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
tensorflow
keras
tensorboard
ipykernel
streamlit
plotly
python-dotenv